minor refactor

+++ src/main/java/se/crafted/chrisb/ecoCreature/managers/ecoConfigManager.java
@@ -273,32 +273,28 @@ public class ecoConfigManager
     {
         FileConfiguration config = new YamlConfiguration();
 
-        try {
-            if (!file.exists()) {
-                file.getParentFile().mkdir();
-                file.createNewFile();
-                InputStream inputStream = plugin.getResource(file.getName());
-                FileOutputStream outputStream = new FileOutputStream(file);
+        if (!file.exists() && file.getParentFile().mkdir() && file.createNewFile()) {
+            InputStream inputStream = plugin.getResource(file.getName());
+            FileOutputStream outputStream = new FileOutputStream(file);
 
-                byte[] buffer = new byte[8192];
-                int length = 0;
-                while ((length = inputStream.read(buffer)) > 0) {
-                    outputStream.write(buffer, 0, length);
-                }
-
-                inputStream.close();
-                outputStream.close();
-
-                ecoCreature.getEcoLogger().info("Default config written to " + file.getName());
+            byte[] buffer = new byte[8192];
+            int length = 0;
+            while ((length = inputStream.read(buffer)) > 0) {
+                outputStream.write(buffer, 0, length);
             }
 
-            config.load(file);
-            config.setDefaults(YamlConfiguration.loadConfiguration(plugin.getResource(file.getName())));
-            config.options().copyDefaults(true);
+            inputStream.close();
+            outputStream.close();
+
+            ecoCreature.getEcoLogger().info("Default config written to " + file.getName());
         }
-        catch (Exception e) {
-            ecoCreature.getEcoLogger().warning("Default config could not be created!");
+        else {
+            ecoCreature.getEcoLogger().severe("Default config could not be created!");
         }
+
+        config.load(file);
+        config.setDefaults(YamlConfiguration.loadConfiguration(plugin.getResource(file.getName())));
+        config.options().copyDefaults(true);
 
         return config;
     }
